chore(style) Update the task
After pushing my first commit for this task, I started checking the solutions
to see if my resolution of the task was correct. I realized that I've forgot
to add the Hex colours instead of the word ones, and also the numbers three and four
should've been under the same tag for the font-size.